move everything to types package and reexport
this might be a good idea. Or not we will see
