+++ Cargo.toml
@@ -2,6 +2,8 @@
 name = "metio"
 version = "0.1.0"
 edition = "2021"
+authors = ["Kevin Stiehl <kevin.stiehl@numericas.de>"]
+description = "The library for interacting with metio compliant systems"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
